refactoring: lombok exclusions

+++ order-microservice/src/main/java/com/wasd/ordermicroservice/data/order/OrderResponse.java
@@ -1,4 +1,4 @@
 package com.wasd.ordermicroservice.data.order;
 
-public record OrderResponse(Long id, OrderState state, Long sellerId, Long categoryId, Long brandId) {
+public record OrderResponse(Long id, OrderState state, Long sellerId, Long categoryId, Integer brandId) {
 }
